feat: Failing to resolve a target name is not a fatal error anymore
In perf module, failing to resolve the true name of the target is
considered as a fatal error and lead to the target not being monitored.

This patch set the target name as the cgroup path relative to the cgroup
basedir in this case and allow the cgroup to be monitored.