Resolve network name for socket output
When the address provided in the configuration is not an IPV4 address
(such as w.x.y.z) we now try to resolve the network name before
connecting.
This will allow using the socket output in environments where IP
addresses are dynamically provided and not known when configuring the
sensor (e.g. docker compose or k8s).