feat: Verify that perf_event works instead of the root user
Initially, the sensor checked if it was run as root before continuing.
However, the sensor can be run as a non-root user if the CAP_PERFMON
or CAP_SYS_ADMIN is granted to the process.
This commit changes the default behavior to a new perf_event_open
self-test which confirms its availability before continuing.