build: Add support for rootless Docker image
This commit adds the powerapi user, set it as default for the image,
and set the file capability to the sensor executable to allow its
rootless usage.

The sensor only needs the CAP_PERFMON capability to work, but it is only
available since Linux 5.8. By default, we set the CAP_SYS_ADMIN capability
that is a *lot* less restrictive but always supported since Linux 2.2.

However, the FILE_CAPABILITY build arg allows the user to change the
default capability when building the image and set it to the more
restricive CAP_PERFMON capability.

The image entrypoint script check if the capability is available in the
current bounding set before starting the sensor. This allows to provide
a meaningful error message to the user instead of the generic EPERM.