fix(Dockerfile): Do not switch to powerapi user by default
For some targets, the name resolution needs to access files owned by
root. Setting the default image user to powerapi breaks the name
resolution for these targets.

This commit removes the switch to the powerapi user by default but keeps
the user availble in the image. This allows user that don't want to run
the container as root to switch to the powerapi user at will.